Cleanup code using forward declarations.
Use misc methods of avoiding unnecesary header includes.
Replace int typedefs with int##_t from stdint.h.
Replace PRI64[xdu] with PRI[xdu]64 from inttypes.h.
Normalize QT_VERSION ifs where possible.
Resolve some indirect dependencies as direct ones.
Remove extern declarations from .cpp files.

Conflicts:
	src/bitcoinrpc.h
	src/checkpoints.cpp
	src/checkpoints.h
	src/core.h
	src/init.cpp
	src/main.cpp
	src/main.h
	src/miner.cpp
	src/miner.h
	src/net.cpp
	src/net.h
	src/qt/intro.cpp
	src/qt/optionsmodel.cpp
	src/qt/sendcoinsdialog.cpp
	src/qt/sendcoinsdialog.h
	src/qt/transactionrecord.cpp
	src/qt/walletframe.cpp
	src/qt/walletmodel.h
	src/rpcmining.cpp
	src/rpcwallet.cpp
	src/test/test_bitcoin.cpp
	src/txdb.h
	src/util.h
	src/wallet.cpp
	src/wallet.h

+++ src/qt/optionsmodel.cpp
@@ -9,12 +9,13 @@
 #include "optionsmodel.h"
 
 #include "bitcoinunits.h"
-#include "init.h"
-#include "core.h"
-#include "wallet.h"
-#include "netbase.h"
-#include "walletdb.h"
 #include "guiutil.h"
+
+#include "init.h"
+#include "main.h"
+#include "net.h"
+#include "wallet.h"
+#include "walletdb.h"
 
 #include <QSettings>
 
@@ -203,7 +204,7 @@ QVariant OptionsModel::data(const QModelIndex & index, int role) const
                 return QVariant(5);
         }
         case Fee:
-            return QVariant(nTransactionFee);
+            return QVariant((qint64) nTransactionFee);
         case SpendZeroConfChange:
             return bSpendZeroConfChange;
         case DisplayUnit:
@@ -281,7 +282,7 @@ bool OptionsModel::setData(const QModelIndex & index, const QVariant & value, in
         break;
         case Fee:
             nTransactionFee = value.toLongLong();
-            settings.setValue("nTransactionFee", nTransactionFee);
+            settings.setValue("nTransactionFee", (qint64) nTransactionFee);
             emit transactionFeeChanged(nTransactionFee);
             break;
         case SpendZeroConfChange:
@@ -319,7 +320,7 @@ bool OptionsModel::setData(const QModelIndex & index, const QVariant & value, in
 
 qint64 OptionsModel::getTransactionFee()
 {
-    return nTransactionFee;
+    return (qint64) nTransactionFee;
 }
 
 bool OptionsModel::getProxySettings(QString& proxyIP, quint16 &proxyPort) const
